+++ processor/src/test/fixtures/bad_input/com/example/fragment/includes/local_only/CrossPackageIncludesModel.java
@@ -1,0 +1,8 @@
+package com.example.fragment.includes.local_only;
+
+import sting.Fragment;
+
+@Fragment( includes = com.example.fragment.includes.local_only.other.MyModel.class )
+public interface CrossPackageIncludesModel
+{
+}