+++ processor/src/test/fixtures/bad_input/com/example/fragment/includes/local_only/other/MyModel.java
@@ -1,0 +1,11 @@
+package com.example.fragment.includes.local_only.other;
+
+import sting.Injectable;
+
+@Injectable
+public class MyModel
+{
+  MyModel()
+  {
+  }
+}