+++ processor/src/test/fixtures/input/com/example/fragment/includes/local_only_disabled/CrossPackageIncludesModel.java
@@ -1,0 +1,8 @@
+package com.example.fragment.includes.local_only_disabled;
+
+import sting.Fragment;
+
+@Fragment( localOnly = false, includes = com.example.fragment.includes.local_only_disabled.other.MyModel.class )
+public interface CrossPackageIncludesModel
+{
+}