+++ processor/src/test/fixtures/input/com/example/fragment/includes/local_only_disabled/other/MyModel.java
@@ -1,0 +1,11 @@
+package com.example.fragment.includes.local_only_disabled.other;
+
+import sting.Injectable;
+
+@Injectable
+public class MyModel
+{
+  MyModel()
+  {
+  }
+}