+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+* Add `@Fragment.localOnly`, default it to `true`, and reject fragment includes declared in a different package unless the flag is explicitly disabled.
 * Update the `org.realityforge.braincheck` artifacts to version `1.32.0`.
 * Add the `sting.warnings_as_errors` annotation processor option to upgrade Sting warnings to errors.
 * Introduce the `@sting.Factory` annotation to generate injectable factory implementations that can be included in